requirements.txt
google-generativeai
google-auth
google-auth-oauthlib
google-api-python-client
python-dotenv
